feat: move event changelog classes into event package [DHIS2-16740]
