+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/event/DefaultEventChangeLogService.java
@@ -25,18 +25,46 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.tracker.export.changelog;
+package org.hisp.dhis.tracker.export.event;
 
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 import org.hisp.dhis.common.UID;
 import org.hisp.dhis.feedback.NotFoundException;
+import org.hisp.dhis.program.Event;
+import org.hisp.dhis.trackedentity.TrackerAccessManager;
+import org.hisp.dhis.user.CurrentUserUtil;
+import org.hisp.dhis.user.User;
+import org.hisp.dhis.user.UserService;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface ChangeLogService {
+@Service("org.hisp.dhis.tracker.export.event.ChangeLogService")
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class DefaultEventChangeLogService implements EventChangeLogService {
 
-  /**
-   * Retrieves the change log data for a particular event
-   *
-   * @return list with the change logs of the supplied event, if any
-   */
-  List<EventChangeLog> getEventChangeLog(UID event) throws NotFoundException;
+  private final org.hisp.dhis.program.EventService eventService;
+
+  private final JdbcEventChangeLogStore jdbcEventChangeLogStore;
+
+  private final TrackerAccessManager trackerAccessManager;
+
+  private final UserService userService;
+
+  @Override
+  public List<EventChangeLog> getEventChangeLog(UID eventUid) throws NotFoundException {
+    Event event = eventService.getEvent(eventUid.getValue());
+    if (event == null) {
+      throw new NotFoundException(Event.class, eventUid.getValue());
+    }
+
+    User currentUser = userService.getUserByUsername(CurrentUserUtil.getCurrentUsername());
+    List<String> errors = trackerAccessManager.canRead(currentUser, event, false);
+    if (!errors.isEmpty()) {
+      throw new NotFoundException(Event.class, eventUid.getValue());
+    }
+
+    return jdbcEventChangeLogStore.getEventChangeLog(eventUid.getValue());
+  }
 }
